Changed the broadcast
Changed the broadcast to exit the DiceFragment. Previously it was done using the handler of said fragment. This ended in the crashing of the game if the fragment has never been initialized. The broadcast now is done from the GameBoardUI and is able to force every client to exit all opened fragments.